Full stack website using python, flask, bootstarp and other technologies
The online auction website is different from others because it uses it's user wallet for the transaction. Secondly, it does notthe  use selinks tok for forgetting passwords so every time user gets a link with an expiration time. User can add product and bid on other products at the same time all all things are working asynchronies so without any lagging the website run continuously. All the data is stored in ORM format and use SQLite database and hashing algorithms to encrypt the password in case someone hacks the database but not gets the user's credentials. Also, it uses to follow and unfollow method to see other users' products as an Instagram posts.